mysite/requirement.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
Deprecated==1.2.14
dj-rest-auth==3.0.0
Django==4.2
django-allauth==0.54.0
django-cors-headers==4.0.0
django-environ==0.10.0
django-oauth-toolkit==2.3.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-social-oauth2==2.1.3
ecdsa==0.18.0
frozenlist==1.3.3
idna==3.4
jwcrypto==1.5.0
multidict==6.0.4
oauthlib==3.2.2
openai==0.27.8
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
PyJWT==2.6.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tqdm==4.65.0
urllib3==1.26.15
wrapt==1.15.0
yarl==1.9.2
